pyproject.toml
[project]
name = "ecommerce-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.1",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
    "djoser>=2.3.3",
    "markdown>=3.10.1",
    "pytest>=9.0.2",
]
